requirements.txt
Pillow 
streamlit
rembg
onnxruntime
ultralytics
torch
torchvision
torchaudio
supervision
opencv-python